refine names of concepts about ritrin point
